# api/openapi.yaml
openapi: "3.1.0"
info:
  title: binge-os-watch API
  description: |
    REST API for binge-os-watch (BINGE) — a fast, opinionated media tracker
    powered by TMDB. Tracks movies and TV shows with per-episode progress,
    ratings at every level, and keyword-based discovery.

    **Authentication**: Login returns a session token and sets an HttpOnly
    session cookie. For browser usage the cookie is used automatically.
    For programmatic access, pass the token as `Authorization: Bearer <token>`.

    **Pagination**: Uses `page_size` and `page_token` query parameters.
    Responses include `next_page_token` (empty when no more pages).

    **Partial updates**: PATCH endpoints accept `update_mask` as a
    comma-separated list of field names to update.

    **Error format**: All errors return `{"code": "ERROR_CODE", "message": "..."}`.

    **Status state machine (TV shows)**:
    - 0 episodes watched → plan_to_watch (auto)
    - 1+ episodes, not all → watching (auto)
    - all episodes watched → watched (auto)
    - user sets on_hold → on_hold (manual, status_override=1)
    - user sets dropped → dropped (manual, status_override=1)
    - user marks episode after override → watching (auto, status_override=0)
  version: "1.0.0"
  license:
    name: AGPL-3.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/users:change-password:
    post:
      tags: [Auth]
      summary: Change the current user's password
      operationId: changePassword
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: Password changed
        "400":
          $ref: "#/components/responses/InvalidArgument"
...
  /api/v1/media/{id}:rate:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Ratings]
      summary: Rate a movie or TV show
      operationId: rateMedia
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRequest"
      responses:
        "204":
          description: Rating saved
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/episodes/{id}:watch:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Episodes]
      summary: Mark an episode as watched
      description: |
        Records watch progress and auto-derives the parent TV show's status.
        Clears any manual status override (on_hold/dropped).
      operationId: watchEpisode
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "204":
          description: Episode marked watched
...
  /api/v1/episodes/{id}:rate:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Ratings]
      summary: Rate an episode
      operationId: rateEpisode
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRequest"
      responses:
        "204":
          description: Rating saved
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/seasons/{id}:watch-all:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Episodes]
      summary: Mark all episodes in a season as watched
      operationId: watchAllInSeason
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "204":
          description: All episodes in season marked watched
  /api/v1/seasons/{id}:rate:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Ratings]
      summary: Rate a season
      operationId: rateSeason
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRequest"
      responses:
        "204":
          description: Rating saved
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Ratings ───────────────────────────────────────────────────────────
  /api/v1/ratings/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    delete:
      tags: [Ratings]
      summary: Delete a rating (legacy, no-op)
      description: |
        Under the Option B schema, deleting a rating is expressed as a score=0
        on the corresponding `:rate` endpoint; the dedicated delete route is
        kept for legacy clients and is a server-side no-op that always returns 204.
      operationId: deleteRating
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "204":
          description: No content (always, even if no rating existed)

  # ── Tags ──────────────────────────────────────────────────────────────
...
  /api/v1/settings:
    get:
      tags: [Settings]
      summary: Get user settings
      operationId: getSettings
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      tags: [Settings]
      summary: Update user settings
      operationId: updateSettings
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/updateMask"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          $ref: "#/components/responses/InvalidArgument"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        HttpOnly session cookie set by the login endpoint.
        SameSite=Strict, Secure configurable.
        30-day sliding expiry (extended on activity).
    bearerToken:
      type: http
      scheme: bearer
      description: |
        Session token returned by the login endpoint. Pass as
        `Authorization: Bearer <token>`. Uses the same session
        as the cookie — no separate API key table.
  parameters:
    id:
      name: id
      in: path
      required: true
      description: |
        Under the Option B schema this is the `user_library.id` of the user's
        library entry, not the TMDB id. The catalog (`tmdb_show` / `tmdb_movie`)
        is shared; user-scoped state lives on `user_library`.
      schema:
        type: string
        format: uuid
...
    updateMask:
      name: update_mask
      in: query
      required: true
      description: Comma-separated list of fields to update
      schema:
        type: string
      example: locale,theme,region
  schemas:
...
    UserSettings:
      type: object
      properties:
        user_id:
          type: string
        locale:
          type: string
          enum: [en, de, nl]
        theme:
          type: string
          enum: [light, dark, oled]
        region:
          type: string
          description: ISO 3166-1 2-letter code for watch provider lookups
          example: NL
        updated_at:
          type: string
          format: date-time
...
    RateRequest:
      type: object
      required: [score]
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 10
...
    AppError:
      type: object
      properties:
        code:
          type: string
          enum:
            - INVALID_ARGUMENT
            - NOT_FOUND
            - ALREADY_EXISTS
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - INTERNAL
        message:
          type: string
        details:
          nullable: true
  responses:
    InvalidArgument:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
          example:
            code: INVALID_ARGUMENT
            message: "title: must not be empty"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppError"
          example:
            code: NOT_FOUND
            message: media not found
...